--- client/src/components/Keywords/fakeData.js
@@ -1,19 +1,0 @@
-const keywordsData = [
-  { text: 'Solo traveller friendly', value: 'solo' },
-  { text: 'Great for nature', value: 'nature' },
-  { text: 'Kids friendly', value: 'kids' },
-  { text: 'Grate for wildlife', value: 'wildlife' },
-  { text: 'Vegetarian friendly', value: 'vegetarian' },
-  { text: 'Great for wellbeing', value: 'wellbeing' },
-  { text: 'Vegan', value: 'vegan' },
-  { text: 'Greate for making new friends', value: 'makingFeiends' },
-  { text: 'Budget', value: 'budget' },
-  { text: 'Great for local culture', value: 'culture' },
-  { text: 'Mid range', value: 'midRange' },
-  { text: 'Great staff', value: 'staff' },
-  { text: 'Luxury', value: 'luxury' },
-  { text: 'Great owners/hosts', value: 'owners' },
-  { text: 'Family owned', value: 'family' },
-];
-
-export default keywordsData;
