--- client/src/components/MagicalFactor/keywords.js
@@ -1,22 +1,0 @@
-const keywordsData = {
-  solo: { text: 'Solo traveller friendly', src: 'solo' },
-  wildlife: { text: 'Great for wildlife', src: 'wild' },
-  family: { text: 'Family owned', src: 'family' },
-  // vegetarian: { text: 'Vegetarian friendly', src: 'vegetarian.png' },
-  // wellbeing: { text: 'Great for wellbeing', src: 'wellbeing.png' },
-  // vegan: { text: 'Vegan', src: 'vegan.png' },
-  // makingFriends: {
-  //   text: 'Great for making new friends',
-  //   src: 'social.png',
-  // },
-  // budget: { text: 'Budget', src: '.png' },
-  // nature: { text: 'Great for nature', src: 'nature.png' },
-  // kids: { text: 'Kids friendly', src: 'kid.png' },
-  // culture: { text: 'Great for local culture', src: '.png' },
-  // midRange: { text: 'Mid range', src: '.png' },
-  // staff: { text: 'Great staff', src: 'staff.png' },
-  // luxury: { text: 'Luxury', src: 'luxury.png' },
-  // owners: { text: 'Great owners/hosts', src: '.png' },
-};
-
-export default keywordsData;
