+++ client/src/staticDataSet/keywords.js
@@ -1,0 +1,22 @@
+const keywordsData = {
+  solo: { text: 'Solo traveller friendly', src: 'solo' },
+  wildlife: { text: 'Great for wildlife', src: 'wild' },
+  family: { text: 'Family owned', src: 'family' },
+  vegetarian: { text: 'Vegetarian friendly', src: 'vegetarian' },
+  wellbeing: { text: 'Great for wellbeing', src: 'wellbeing' },
+  vegan: { text: 'Vegan', src: 'vegan' },
+  makingFriends: {
+    text: 'Great for making new friends',
+    src: 'social',
+  },
+  budget: { text: 'Budget', src: '' },
+  nature: { text: 'Great for nature', src: 'nature' },
+  kids: { text: 'Kids friendly', src: 'kid' },
+  culture: { text: 'Great for local culture', src: '' },
+  midRange: { text: 'Mid range', src: '' },
+  staff: { text: 'Great staff', src: 'staff' },
+  luxury: { text: 'Luxury', src: 'luxury' },
+  owners: { text: 'Great owners/hosts', src: '' },
+};
+
+export default keywordsData;
